tagger/requirements.txt
boto3==1.18.26
botocore==1.21.26
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
flake8==3.9.2
google-api-core==2.0.0
google-api-python-client==2.17.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jmespath==0.10.0
mccabe==0.6.1
oauthlib==3.1.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
uritemplate==3.0.1
urllib3==1.26.6
